# arcx_backend/openapi.yaml
openapi: 3.0.3
info:
  title: ARCX Platform API
  version: 1.0.0
  description: |-
    ## ARCX — Algorithmic Reserve Currency Exchange

    ARCX is an INR-denominated, NAV-backed digital asset platform. Investors deposit INR which is converted to ARCX tokens at the current NAV price. The vault holds a diversified portfolio (SPY, TLT, GLD) rebalanced algorithmically.

    ### Authentication
    All protected endpoints require a **Bearer JWT** in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```

    Obtain tokens via `POST /api/auth/token/` or `POST /api/v1/auth/register`.

    ### Idempotency
    Financial mutation endpoints (deposit, withdraw, transfer) require an `Idempotency-Key` UUID header to prevent duplicate transactions.

    ### Error Format
    All errors follow: `{"error": "message", "code": "SNAKE_CASE_CODE"}`
  contact:
    name: ARCX Engineering
    email: engineering@arcx.in
  license:
    name: Proprietary
paths:
...
  /api/v1/wallet/:
    get:
      operationId: wallet_balance
      description: "Returns the authenticated user's ARCX token balance, cost basis,
        and real-time unrealized P&L (calculated using the current oracle NAV). \n\n`unrealized_pnl_inr`
        may return `null` if the oracle is temporarily unavailable — the balance is
        always returned."
      summary: Get wallet balance and P&L
      tags:
      - Wallet
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorResponse'
          description: JWT missing or expired
  /api/v1/wallet/deposit:
    post:
      operationId: wallet_deposit
      description: |-
        Converts INR to ARCX tokens at the current oracle NAV price. Minimum deposit is ₹100. KYC approval is required.

        **Idempotency:** Include a unique `Idempotency-Key` UUID header. Replaying the same key returns the original response without double-charging.
      summary: Deposit INR -> ARCX
      parameters:
      - in: header
        name: Idempotency-Key
        schema:
          type: string
          format: uuid
        description: A unique UUID v4 per request. Prevents duplicate transactions
          on retry.
        required: true
      tags:
      - Wallet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequestRequest'
            examples:
              Deposit5000INR:
                value:
                  amount_inr: '5000.00'
                summary: Deposit 5000 INR
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DepositRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DepositRequestRequest'
        required: true
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositResponse'
              examples:
                DepositSuccess:
                  value:
                    transaction_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    arcx_credited: '49.123456789012345678'
                    nav_at_tx: '101.80'
                    new_balance: '49.123456789012345678'
                  summary: Deposit success
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorResponse'
          description: Validation error or oracle failure
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorResponse'
          description: KYC not approved or circuit breaker active
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorResponse'
          description: JWT missing or expired
  /api/v1/wallet/withdraw:
    post:
      operationId: wallet_withdraw
      description: |-
        Redeems ARCX tokens back to INR at the current oracle NAV price. A redemption fee is deducted — see `fee_inr` in the response. KYC approval and an active (non-frozen) wallet are both required.

        **Idempotency:** Include `Idempotency-Key` UUID header to prevent double-redemptions.
      summary: Withdraw ARCX -> INR
      parameters:
      - in: header
        name: Idempotency-Key
        schema:
          type: string
          format: uuid
        description: A unique UUID v4 per request. Prevents duplicate redemptions
          on retry.
        required: true
      tags:
      - Wallet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawRequestRequest'
            examples:
              Withdraw10.5ARCX:
                value:
                  amount_arcx: '10.5'
                summary: Withdraw 10.5 ARCX
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/WithdrawRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/WithdrawRequestRequest'
        required: true
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawResponse'
              examples:
                WithdrawSuccess:
                  value:
                    transaction_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    gross_inr: '1069.25'
                    fee_inr: '0.35'
                    inr_returned: '1068.90'
                    nav_at_tx: '101.80'
                    new_balance: '38.623456789012345678'
                  summary: Withdraw success
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorResponse'
          description: Insufficient balance or validation error
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorResponse'
          description: KYC not approved, wallet frozen, or circuit breaker active
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorResponse'
          description: JWT missing or expired
  /api/v1/wallet/history:
    get:
      operationId: wallet_history
      description: Returns the most recent N transactions for the authenticated user,
        ordered newest-first. Includes deposits, withdrawals, transfers, and dividends.
        Maximum `limit` is capped at 100.
      summary: Get transaction history
      parameters:
      - in: query
        name: limit
        schema:
          type: integer
          default: 20
        description: Number of transactions to return. Default 20, max 100.
      tags:
      - Wallet
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorResponse'
          description: JWT missing or expired
...
components:
  schemas:
...
    DepositRequestRequest:
      type: object
      description: |-
        Validates a deposit request.

        amount_inr: How much INR the user wants to convert to ARCX.
        Minimum ₹100. No maximum (KYC tier limit enforced in service layer).
      properties:
        amount_inr:
          type: string
          format: decimal
          pattern: ^-?\d{0,18}(?:\.\d{0,2})?$
      required:
      - amount_inr
    DepositResponse:
      type: object
      properties:
        transaction_id:
          type: string
          format: uuid
          description: Unique transaction ID
        arcx_credited:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,18})?$
          description: ARCX tokens credited to wallet
        nav_at_tx:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,4})?$
          description: NAV price used for this conversion (INR)
        new_balance:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,18})?$
          description: Updated ARCX wallet balance
      required:
      - arcx_credited
      - nav_at_tx
      - new_balance
      - transaction_id
...
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            type: object
            additionalProperties: {}
          description: List of transaction records (id, tx_type, amount_arcx, amount_inr,
            fee_inr, nav_at_tx, status, created_at)
        count:
          type: integer
      required:
      - count
      - transactions
...
    WalletErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
      - code
      - error
    WalletResponse:
      type: object
      description: |-
        Shapes wallet data for the response.
        Never exposes internal fields like is_frozen or cost_basis directly.
        unrealized_pnl_inr is a computed field — not stored in DB.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        user_email:
          type: string
          format: email
          readOnly: true
        arcx_balance:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,18})?$
        cost_basis_inr:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,4})?$
        unrealized_pnl_inr:
          type: string
          description: |-
            P&L = (current_value) - (what_you_paid)
            current_value = arcx_balance * current_nav_inr

            current_nav_inr is injected via serializer context from the view.
            If not available (e.g., oracle down), returns None gracefully.
            Returns a decimal string or null.
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - id
      - unrealized_pnl_inr
      - updated_at
      - user_email
    WithdrawRequestRequest:
      type: object
      description: |-
        Validates a withdrawal request.

        amount_arcx: How many ARCX tokens to sell back.
        Minimum 0.01 ARCX.
      properties:
        amount_arcx:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,18})?$
      required:
      - amount_arcx
    WithdrawResponse:
      type: object
      properties:
        transaction_id:
          type: string
          format: uuid
        gross_inr:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,4})?$
          description: INR before fees
        fee_inr:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,4})?$
          description: Redemption fee charged
        inr_returned:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,4})?$
          description: Net INR returned to user
        nav_at_tx:
          type: string
          format: decimal
          pattern: ^-?\d{0,16}(?:\.\d{0,4})?$
          description: NAV price used for this conversion (INR)
        new_balance:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,18})?$
          description: Updated ARCX wallet balance
      required:
      - fee_inr
      - gross_inr
      - inr_returned
      - nav_at_tx
      - new_balance
      - transaction_id
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...